Separated user criteria view into an element
